Revert "Style Book: Add color tab (#65692)"
This reverts commit 54371146e6515a15118c6feceb34ade061b44a1b.
