+++ packages/edit-site/src/components/style-book/index.js
@@ -11,26 +11,18 @@ import {
 	Composite,
 	privateApis as componentsPrivateApis,
 } from '@wordpress/components';
-import { __, _x, sprintf } from '@wordpress/i18n';
+import { __, sprintf } from '@wordpress/i18n';
 import {
 	BlockList,
 	privateApis as blockEditorPrivateApis,
 	store as blockEditorStore,
-	useSettings,
 	__unstableEditorStyles as EditorStyles,
 	__unstableIframe as Iframe,
-	__experimentalUseMultipleOriginColorsAndGradients as useMultipleOriginColorsAndGradients,
 } from '@wordpress/block-editor';
 import { privateApis as editorPrivateApis } from '@wordpress/editor';
 import { useSelect } from '@wordpress/data';
 import { useResizeObserver } from '@wordpress/compose';
-import {
-	useMemo,
-	useState,
-	memo,
-	useContext,
-	useEffect,
-} from '@wordpress/element';
+import { useMemo, useState, memo, useContext } from '@wordpress/element';
 import { ENTER, SPACE } from '@wordpress/keycodes';
 
 /**
@@ -59,81 +51,6 @@ function isObjectEmpty( object ) {
 	return ! object || Object.keys( object ).length === 0;
 }
 
-/**
- * Retrieves colors, gradients, and duotone filters from Global Styles.
- * The inclusion of default (Core) palettes is controlled by the relevant
- * theme.json property e.g. defaultPalette, defaultGradients, defaultDuotone.
- *
- * @return {Object} Object containing properties for each type of palette.
- */
-function useMultiOriginPalettes() {
-	const { colors, gradients } = useMultipleOriginColorsAndGradients();
-
-	// Add duotone filters to the palettes data.
-	const [
-		shouldDisplayDefaultDuotones,
-		customDuotones,
-		themeDuotones,
-		defaultDuotones,
-	] = useSettings(
-		'color.defaultDuotone',
-		'color.duotone.custom',
-		'color.duotone.theme',
-		'color.duotone.default'
-	);
-
-	const palettes = useMemo( () => {
-		const result = { colors, gradients, duotones: [] };
-
-		if ( themeDuotones && themeDuotones.length ) {
-			result.duotones.push( {
-				name: _x(
-					'Theme',
-					'Indicates these duotone filters come from the theme.'
-				),
-				slug: 'theme',
-				duotones: themeDuotones,
-			} );
-		}
-
-		if (
-			shouldDisplayDefaultDuotones &&
-			defaultDuotones &&
-			defaultDuotones.length
-		) {
-			result.duotones.push( {
-				name: _x(
-					'Default',
-					'Indicates these duotone filters come from WordPress.'
-				),
-				slug: 'default',
-				duotones: defaultDuotones,
-			} );
-		}
-		if ( customDuotones && customDuotones.length ) {
-			result.duotones.push( {
-				name: _x(
-					'Custom',
-					'Indicates these doutone filters are created by the user.'
-				),
-				slug: 'custom',
-				duotones: customDuotones,
-			} );
-		}
-
-		return result;
-	}, [
-		colors,
-		gradients,
-		customDuotones,
-		themeDuotones,
-		defaultDuotones,
-		shouldDisplayDefaultDuotones,
-	] );
-
-	return palettes;
-}
-
 function StyleBook( {
 	enableResizing = true,
 	isSelected,
@@ -147,8 +64,7 @@ function StyleBook( {
 	const [ resizeObserver, sizes ] = useResizeObserver();
 	const [ textColor ] = useGlobalStyle( 'color.text' );
 	const [ backgroundColor ] = useGlobalStyle( 'color.background' );
-	const colors = useMultiOriginPalettes();
-	const [ examples, setExamples ] = useState( () => getExamples( colors ) );
+	const [ examples ] = useState( getExamples );
 	const tabs = useMemo(
 		() =>
 			getTopLevelStyleBookCategories().filter( ( category ) =>
@@ -158,12 +74,6 @@ function StyleBook( {
 			),
 		[ examples ]
 	);
-
-	// Ensure color examples are kept in sync with Global Styles palette changes.
-	useEffect( () => {
-		setExamples( getExamples( colors ) );
-	}, [ colors ] );
-
 	const { base: baseConfig } = useContext( GlobalStylesContext );
 
 	const mergedConfig = useMemo( () => {
@@ -364,7 +274,6 @@ const Examples = memo(
 							key={ example.name }
 							id={ `example-${ example.name }` }
 							title={ example.title }
-							content={ example.content }
 							blocks={ example.blocks }
 							isSelected={ isSelected( example.name ) }
 							onClick={ () => {
@@ -403,7 +312,6 @@ const Subcategory = ( { examples, isSelected, onSelect } ) => {
 				key={ example.name }
 				id={ `example-${ example.name }` }
 				title={ example.title }
-				content={ example.content }
 				blocks={ example.blocks }
 				isSelected={ isSelected( example.name ) }
 				onClick={ () => {
@@ -414,9 +322,7 @@ const Subcategory = ( { examples, isSelected, onSelect } ) => {
 	);
 };
 
-const disabledExamples = [ 'example-duotones' ];
-
-const Example = ( { id, title, blocks, isSelected, onClick, content } ) => {
+const Example = ( { id, title, blocks, isSelected, onClick } ) => {
 	const originalSettings = useSelect(
 		( select ) => select( blockEditorStore ).getSettings(),
 		[]
@@ -436,20 +342,12 @@ const Example = ( { id, title, blocks, isSelected, onClick, content } ) => {
 		[ blocks ]
 	);
 
-	const disabledProps = disabledExamples.includes( id )
-		? {
-				disabled: true,
-				accessibleWhenDisabled: true,
-		  }
-		: {};
-
 	return (
 		<div role="row">
 			<div role="gridcell">
 				<Composite.Item
 					className={ clsx( 'edit-site-style-book__example', {
 						'is-selected': isSelected,
-						'is-disabled-example': !! disabledProps?.disabled,
 					} ) }
 					id={ id }
 					aria-label={ sprintf(
@@ -460,7 +358,6 @@ const Example = ( { id, title, blocks, isSelected, onClick, content } ) => {
 					render={ <div /> }
 					role="button"
 					onClick={ onClick }
-					{ ...disabledProps }
 				>
 					<span className="edit-site-style-book__example-title">
 						{ title }
@@ -470,17 +367,12 @@ const Example = ( { id, title, blocks, isSelected, onClick, content } ) => {
 						aria-hidden
 					>
 						<Disabled className="edit-site-style-book__example-preview__content">
-							{ content ? (
-								content
-							) : (
-								<ExperimentalBlockEditorProvider
-									value={ renderedBlocks }
-									settings={ settings }
-								>
-									<EditorStyles />
-									<BlockList renderAppender={ false } />
-								</ExperimentalBlockEditorProvider>
-							) }
+							<ExperimentalBlockEditorProvider
+								value={ renderedBlocks }
+								settings={ settings }
+							>
+								<BlockList renderAppender={ false } />
+							</ExperimentalBlockEditorProvider>
 						</Disabled>
 					</div>
 				</Composite.Item>
