+++ packages/edit-site/src/components/style-book/types.ts
@@ -9,7 +9,6 @@ export type StyleBookCategory = {
 	slug: string;
 	blocks?: string[];
 	exclude?: string[];
-	include?: string[];
 	subcategories?: StyleBookCategory[];
 };
 
@@ -17,8 +16,7 @@ export type BlockExample = {
 	name: string;
 	title: string;
 	category: string;
-	content?: JSX.Element;
-	blocks?: Block | Block[];
+	blocks: Block | Block[];
 };
 
 export type CategoryExamples = {
@@ -26,35 +24,4 @@ export type CategoryExamples = {
 	slug: string;
 	examples?: BlockExample[];
 	subcategories?: CategoryExamples[];
-};
-
-export type StyleBookColorGroup = {
-	origin: string;
-	slug: string;
-	title: string;
-	type: string;
-};
-
-export type Color = { slug: string };
-export type Gradient = { slug: string };
-export type Duotone = {
-	colors: string[];
-	slug: string;
-};
-
-export type ColorOrigin = {
-	name: string;
-	slug: string;
-	colors?: Color[];
-	gradients?: Gradient[];
-	duotones?: Duotone[];
-};
-
-export type MultiOriginPalettes = {
-	disableCustomColors: boolean;
-	disableCustomGradients: boolean;
-	hasColorsOrGradients: boolean;
-	colors: Omit< ColorOrigin, 'gradients' | 'duotones' >;
-	duotones: Omit< ColorOrigin, 'colors' | 'gradients' >;
-	gradients: Omit< ColorOrigin, 'colors' | 'duotones' >;
 };
